Make /etc/default/mythtv-status a bit clearer and fix a bug with EMAIL_ARGS in the postinst.
The comments in /etc/default/mythtv-status should now be a bit clearer about
which scripts use which variables.

In the postinst I was setting EMAIL_ARGS to a default setting it wasn't set.
This caused it not "remember" if someone had removed the arguments and always
reset it to the default setting.  Now the default setting is set before the
config file is loaded.  So that any settings (or unsettings) in the config
file will be used.