The XML since r20037 contains some additional states, cater for them.
From Marc:

OK, not good.

From my research, I am on ProtoVer 44 which was switched at 19978.
The changes to State went in at 20037, well after the switch to 44.
I am currently on 20281.

Not looking like we can use ProtoVer as the versioning check atm.

Marc

So we'll pretend that protocol version 44 is when the change happened.