Fix the dependencies in the init.d file.
With dependency based boot sequencing, I discovered what I believe is
a bug in the init.d scripts.  The script uses files in /usr/, and
should depend on $remote_fs.  Further, the script depend on the
non-existing facility $mythtv-backend.  This do not exist in the
Debian archive as far as I can see, and I suspect it should be some
facility provided by the mythtv scripts.  I suggest changing this to
not be a virtual facility and to make the dependency optional to avoid
being unable to install the script when the facility is not available.
The stop dependencies should probably be the same as the start
facilities.

<URL: http://refspecs.freestandards.org/LSB_2.1.0/LSB-generic/LSB-generic/initscrcomconv.html >
documents the LSB header format.  Some debian notes are available from
<URL: http://wiki.debian.org/LSBInitScripts >.