Fix passing in the error ref to DateCalc.
It seems that Date::Manip::DateCalc in some versions require that
you parse in a ref for the error ref, passing in undef causes it
to not perform the date calculation.  Parse in a ref, even if we
don't actually check it.  Thanks to Ron Kellam for doing some
leg work on tracking this issue down.